WIP Tests in CI
