+++ melatonin/cached_blur.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "support/implementations.h"
+#include "support/helpers.h"
 
 namespace melatonin
 {
