+++ melatonin/implementations/vImage.h
@@ -1,6 +1,6 @@
 #pragma once
-#include "juce_gui_basics/juce_gui_basics.h"
 #include "Accelerate/Accelerate.h"
+#include "juce_gui_basics/juce_gui_basics.h"
 
 namespace melatonin::blur
 {
@@ -64,5 +64,26 @@ namespace melatonin::blur
         // BGRA = little endian ARGB
         vImagePremultiplyData_BGRA8888 (&dstBuffer, &dstBuffer, kvImageNoFlags);
         return dst;
+    }
+
+    static void tentBlurSingleChannel (juce::Image& img, unsigned int radius)
+    {
+        const unsigned int w = (unsigned int) img.getWidth();
+        const unsigned int h = (unsigned int) img.getHeight();
+
+        juce::Image::BitmapData data (img, juce::Image::BitmapData::readWrite);
+
+        vImage_Buffer src = { data.getLinePointer (0), h, w, (size_t) data.lineStride };
+        vImage_Buffer dst = { data.getLinePointer (0), h, w, (size_t) data.lineStride };
+        vImageTentConvolve_Planar8 (
+            &src,
+            &dst,
+            nullptr,
+            0,
+            0,
+            radius * 2 + 1,
+            radius * 2 + 1,
+            0,
+            kvImageEdgeExtend);
     }
 }
