+++ melatonin/implementations/vector.h
@@ -1,6 +1,5 @@
 #pragma once
 #include "juce_gui_basics/juce_gui_basics.h"
-#include "melatonin_perfetto/melatonin_perfetto.h"
 #include "melatonin_vector/melatonin_vector.h"
 
 namespace melatonin::stackBlur
