+++ melatonin/implementations/vector_class.h
@@ -1,6 +1,5 @@
 #pragma once
 #include "juce_gui_basics/juce_gui_basics.h"
-#include "melatonin_perfetto/melatonin_perfetto.h"
 #include "melatonin_vector/melatonin_vector.h"
 #include "melatonin_vector/melatonin/utilities.h"
 
@@ -145,7 +144,6 @@ namespace melatonin
         // This one is optimized for readability and simplicity
         void singleChannel ()
         {
-            TRACE_COMPONENT ("dimensions", img.getWidth(), "radius", radius);
 
             stackBlurPass<Orientation::Horizontal> ();
             stackBlurPass<Orientation::Vertical> ();
@@ -189,7 +187,6 @@ namespace melatonin
 
         inline void reset ()
         {
-            TRACE_COMPONENT();
 
             // queue is just written to, but the rest are added to
             vector::fill (stackSumVector, 0.0f);
@@ -202,8 +199,7 @@ namespace melatonin
         template <Orientation orientation>
         static inline void convertToFloats (juce::Image::BitmapData& data, size_t pixelOffset, std::vector<float>& destination, size_t vectorSize)
         {
-            TRACE_COMPONENT();
-            if constexpr (orientation == Orientation::Horizontal)
+                        if constexpr (orientation == Orientation::Horizontal)
             {
                 vector::convertToFloats (data.getLinePointer (0) + (unsigned int) data.pixelStride * pixelOffset, data.lineStride, destination, vectorSize);
             }
@@ -216,8 +212,7 @@ namespace melatonin
         template <Orientation orientation>
         static inline void convertToUInt8s (juce::Image::BitmapData& data, size_t offset, std::vector<float>& source, size_t vectorSize)
         {
-            TRACE_COMPONENT();
-            if constexpr (orientation == Orientation::Horizontal)
+                        if constexpr (orientation == Orientation::Horizontal)
             {
                 vector::convertToUInt8s (source, data.getLinePointer (0) + (unsigned int) data.pixelStride * offset, data.lineStride, vectorSize);
             }
