+++ melatonin/implementations/vector_optimized.h
@@ -1,6 +1,5 @@
 #pragma once
 #include "juce_gui_basics/juce_gui_basics.h"
-#include "melatonin_perfetto/melatonin_perfetto.h"
 #include "melatonin_vector/melatonin_vector.h"
 
 namespace melatonin::stackBlur
@@ -11,8 +10,7 @@ namespace melatonin::stackBlur
     template <Orientation orientation>
     static inline void convertToFloats (juce::Image::BitmapData& data, size_t batchOffset, size_t pixelOffset, std::vector<float>& destination, size_t vectorSize)
     {
-        TRACE_COMPONENT();
-        if constexpr (orientation == Orientation::Horizontal)
+                if constexpr (orientation == Orientation::Horizontal)
         {
             vector::convertToFloats (data.getLinePointer (batchOffset) + (unsigned int) data.pixelStride * pixelOffset, data.lineStride, destination, vectorSize);
         }
@@ -25,8 +23,7 @@ namespace melatonin::stackBlur
     template <Orientation orientation>
     static inline void convertToUInt8s (juce::Image::BitmapData& data, size_t batchOffset, size_t offset, std::vector<float>& source, size_t vectorSize)
     {
-        TRACE_COMPONENT();
-        if constexpr (orientation == Orientation::Horizontal)
+                if constexpr (orientation == Orientation::Horizontal)
         {
             vector::convertToUInt8s (source, data.getLinePointer (batchOffset) + (unsigned int) data.pixelStride * offset, data.lineStride, vectorSize);
         }
@@ -196,7 +193,6 @@ namespace melatonin::stackBlur
     // This one is optimized for readability and simplicity
     static void vectorOptimizedSingleChannel (juce::Image& img, unsigned int radius)
     {
-        TRACE_COMPONENT ("dimensions", img.getWidth(), "radius", radius);
 
         juce::Image::BitmapData data (img, juce::Image::BitmapData::readWrite);
