+++ melatonin/support/helpers.h
@@ -1,4 +1,6 @@
 #pragma once
+#include "juce_gui_basics/juce_gui_basics.h"
+#include "implementations.h"
 
 namespace melatonin
 {
