+++ melatonin_blur.h
@@ -3,13 +3,14 @@
 /*
 BEGIN_JUCE_MODULE_DECLARATION
 
- ID:               melatonin_blur
- vendor:           Sudara
- version:          1.0.0
- name:             Optimized CPU vector blurring and JUCE drop shadowing with tests and benchmarks
- description:      Blurry Life
- license:          MIT
- dependencies:     juce_graphics
+ID:               melatonin_blur
+vendor:           Sudara
+version:          1.0.0
+name:             Optimized CPU vector blurring and JUCE drop shadowing with tests and benchmarks
+description:      Blurry Life
+license:          MIT
+minimumCppStandard: 17
+dependencies:     juce_graphics,juce_gui_basics,juce_audio_basics
 
 END_JUCE_MODULE_DECLARATION
 */
